Adding synthetic data-generation for Movielens
