+++ merlin/models/data/entertainment/movielens/25m/schema.pbtxt
@@ -1,0 +1,110 @@
+feature {
+  name: "movieId"
+  type: INT
+  int_domain {
+    name: "movieId"
+    max: 56719
+    is_categorical: true
+  }
+  annotation {
+    tag: "item_id"
+    tag: "categorical"
+    tag: "item"
+    extra_metadata {
+      type_url: "type.googleapis.com/google.protobuf.Struct"
+      value: "\n\025\n\010max_size\022\t\021\000\000\000\000\000\000\000\000\nG\n\017embedding_sizes\0224*2\n\026\n\tdimension\022\t\021\000\000\000\000\000\000\200@\n\030\n\013cardinality\022\t\021\000\000\000\000\340\261\353@\n2\n\010cat_path\022&\032$.//categories/unique.movieId.parquet\n\021\n\013num_buckets\022\002\010\000\n\034\n\017dtype_item_size\022\t\021\000\000\000\000\000\000@@\n\030\n\013start_index\022\t\021\000\000\000\000\000\000\000\000\n\017\n\tis_ragged\022\002 \000\n\033\n\016freq_threshold\022\t\021\000\000\000\000\000\000\000\000\n\r\n\007is_list\022\002 \000"
+    }
+  }
+}
+feature {
+  name: "userId"
+  type: INT
+  int_domain {
+    name: "userId"
+    max: 162541
+    is_categorical: true
+  }
+  annotation {
+    tag: "user_id"
+    tag: "user"
+    tag: "categorical"
+    extra_metadata {
+      type_url: "type.googleapis.com/google.protobuf.Struct"
+      value: "\nG\n\017embedding_sizes\0224*2\n\030\n\013cardinality\022\t\021\000\000\000\000h\327\003A\n\026\n\tdimension\022\t\021\000\000\000\000\000\000\200@\n\r\n\007is_list\022\002 \000\n\021\n\013num_buckets\022\002\010\000\n\025\n\010max_size\022\t\021\000\000\000\000\000\000\000\000\n1\n\010cat_path\022%\032#.//categories/unique.userId.parquet\n\033\n\016freq_threshold\022\t\021\000\000\000\000\000\000\000\000\n\034\n\017dtype_item_size\022\t\021\000\000\000\000\000\000@@\n\030\n\013start_index\022\t\021\000\000\000\000\000\000\000\000\n\017\n\tis_ragged\022\002 \000"
+    }
+  }
+}
+feature {
+  name: "genres"
+  value_count {
+  }
+  type: INT
+  int_domain {
+    name: "genres"
+    max: 20
+    is_categorical: true
+  }
+  annotation {
+    tag: "categorical"
+    tag: "item"
+    extra_metadata {
+      type_url: "type.googleapis.com/google.protobuf.Struct"
+      value: "\n\033\n\016freq_threshold\022\t\021\000\000\000\000\000\000\000\000\n\021\n\013num_buckets\022\002\010\000\n\034\n\017dtype_item_size\022\t\021\000\000\000\000\000\000@@\n\r\n\007is_list\022\002 \001\n\030\n\013start_index\022\t\021\000\000\000\000\000\000\000\000\n\017\n\tis_ragged\022\002 \001\n\025\n\010max_size\022\t\021\000\000\000\000\000\000\000\000\nG\n\017embedding_sizes\0224*2\n\030\n\013cardinality\022\t\021\000\000\000\000\000\0004@\n\026\n\tdimension\022\t\021\000\000\000\000\000\0000@\n1\n\010cat_path\022%\032#.//categories/unique.genres.parquet"
+    }
+  }
+}
+feature {
+  name: "TE_movieId_rating"
+  type: FLOAT
+  annotation {
+    tag: "continuous"
+    extra_metadata {
+      type_url: "type.googleapis.com/google.protobuf.Struct"
+      value: "\n\034\n\017dtype_item_size\022\t\021\000\000\000\000\000\000P@\n\017\n\tis_ragged\022\002 \000\n\r\n\007is_list\022\002 \000"
+    }
+  }
+}
+feature {
+  name: "userId_count"
+  type: FLOAT
+  annotation {
+    tag: "continuous"
+    extra_metadata {
+      type_url: "type.googleapis.com/google.protobuf.Struct"
+      value: "\n\017\n\tis_ragged\022\002 \000\n\r\n\007is_list\022\002 \000\n\034\n\017dtype_item_size\022\t\021\000\000\000\000\000\000@@"
+    }
+  }
+}
+feature {
+  name: "rating"
+  type: FLOAT
+  annotation {
+    tag: "regression"
+    tag: "target"
+    extra_metadata {
+      type_url: "type.googleapis.com/google.protobuf.Struct"
+      value: "\n\r\n\007is_list\022\002 \000\n\017\n\tis_ragged\022\002 \000\n\034\n\017dtype_item_size\022\t\021\000\000\000\000\000\000@@"
+    }
+  }
+}
+feature {
+  name: "rating_binary"
+  type: INT
+  annotation {
+    tag: "binary_classification"
+    tag: "target"
+    extra_metadata {
+      type_url: "type.googleapis.com/google.protobuf.Struct"
+      value: "\n\017\n\tis_ragged\022\002 \000\n\034\n\017dtype_item_size\022\t\021\000\000\000\000\000\000@@\n\r\n\007is_list\022\002 \000"
+    }
+  }
+}
+feature {
+  name: "title"
+  annotation {
+    extra_metadata {
+      type_url: "type.googleapis.com/google.protobuf.Struct"
+      value: "\n\034\n\017dtype_item_size\022\t\021\000\000\000\000\000\000P@\n\017\n\tis_ragged\022\002 \000\n\r\n\007is_list\022\002 \000"
+    }
+  }
+}
