+++ merlin/models/data/synthetic.py
@@ -33,10 +33,13 @@ HERE = pathlib.Path(__file__).parent
 KNOWN_DATASETS: Dict[str, Path] = {
     "e-commerce": HERE / "ecommerce/small",
     "e-commerce-large": HERE / "ecommerce/large",
-    "music-streaming": HERE / "music_streaming",
+    "music-streaming": HERE / "entertainment" / "music_streaming",
     "social": HERE / "social",
     "testing": HERE / "testing",
-    "sequence-testing": HERE / "sequence_testing",
+    "sequence-testing": HERE / "testing" / "sequence_testing",
+    "movielens-25m": HERE / "entertainment/movielens/25m",
+    "movielens-1m": HERE / "entertainment/movielens/1m",
+    "movielens-100k": HERE / "entertainment/movielens/100k",
     "criteo": HERE / "advertising/criteo/transformed",
     "aliccp": HERE / "ecommerce/aliccp/transformed",
     "aliccp-raw": HERE / "ecommerce/aliccp/raw",
